refactor transforms, index name inference, search capture mechanism, fix capture search tests
